requirements.txt
Django>=2.2.1,<2.3.0
Pillow==6.0.0

djangorestframework>=3.9.3,<3.10.0
Markdown>=3.1,<3.2
django-filter>=2.1.0,<2.2.0
djangorestframework-jwt==1.11.0

psycopg2>=2.7.5,<2.8.0
psycopg2-binary>=2.8.2,<2.9.0